Adding support for s390x architecture

+++ src/linux/ldcache.cpp
@@ -114,16 +114,14 @@ struct HeaderNew
                           // and align to 8 byte boundary.
 };
 
-
 struct EntryNew
 {
-  int16_t flags;        // Flags bits determine arch and library type.
+  int32_t flags;        // Flags bits determine arch and library type.
   uint32_t key;         // String table index.
   uint32_t value;       // String table index.
   uint32_t osVersion;   // Required OS version.
   uint64_t hwcap;       // Hwcap entry.
 };
-
 
 // Returns a 'boundary' aligned pointer by rounding up to
 // the nearest multiple of 'boundary'.
